refactor, fix consent text appearance
